+++ src/controllers/ProfileResource.php
@@ -2,7 +2,12 @@
 
 namespace roaresearch\yii2\roa\controllers;
 
-use roaresearch\yii2\roa\actions\{ProfileUpdate, ProfileView};
+use roaresearch\yii2\roa\actions\{
+    ProfileFileStream,
+    ProfileUpdate,
+    ProfileView
+};
+use Yii;
 use yii\{base\Model, filters\VerbFilter, rest\OptionsAction};
 
 class ProfileResource extends \yii\rest\Controller
@@ -53,6 +58,7 @@ class ProfileResource extends \yii\rest\Controller
                 'scenario' => $this->updateScenario,
             ],
             'options' => ['class' => OptionsAction::class],
+            'file-stream' => ['class' => ProfileFileStream::class],
         ];
     }
 }